ns for fx: clean up manual string names of related ops
Summary:

Removes the manually specified string name for sets of
related ops, and replaces it with an automatically generated
index. The manual name was arbitrary and ok for an MVP, but
is not safe for wide usage.

Also, adds APIs for users to add custom functions to the
relatedness map by either pairing it to a known function
or creating a new relatedness set.

Test Plan:

```
python test/test_quantization.py TestFXGraphMatcher
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: b3f181285a519a05158e073d0f8114044edf89b5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57210